Feature/replace words

+++ audac_parse_test/prod/output.py
@@ -15,14 +15,29 @@ def output(upd_table, page_name):
     file_path = results_dir / file_name
     df = pd.DataFrame(upd_table)
     sheet_name = page_name
-    with ExcelWriter(
-        file_path,
-        mode='a' if file_path.exists() else 'w'
-    ) as writer:
-        df.to_excel(
-            writer,
-            index=0,
-            header=False,
-            sheet_name=sheet_name
-        )
-    logging.info(f'Файл с результатами был сохранён: {file_path}')
+    if file_path.exists():
+        with ExcelWriter(
+            file_path,
+            mode='a',
+            if_sheet_exists='replace'
+        ) as writer:
+            df.to_excel(
+                writer,
+                index=0,
+                header=False,
+                sheet_name=sheet_name
+            )
+        logging.info(f'Файл с результатами был ПЕРЕЗАПИСАН: {file_path}')
+
+    else:
+        with ExcelWriter(
+            file_path,
+            mode='w',
+        ) as writer:
+            df.to_excel(
+                writer,
+                index=0,
+                header=False,
+                sheet_name=sheet_name
+            )
+        logging.info(f'Файл с результатами был СОХРАНЕН: {file_path}')
